fix: error and analytics handling
