+++ packages/blade-mcp/src/tools/createBladeCursorRules.ts
@@ -2,8 +2,13 @@ import { join } from 'path';
 import { existsSync, unlinkSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
 import type { ToolCallback } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { z } from 'zod';
-import { BLADE_CURSOR_RULES_FILE_PATH, hasOutDatedRules, CURSOR_RULES_VERSION } from '../utils.js';
-import { sendAnalytics } from '../sendAnalytics.js';
+import {
+  BLADE_CURSOR_RULES_FILE_PATH,
+  hasOutDatedRules,
+  CURSOR_RULES_VERSION,
+  handleError,
+  sendAnalytics,
+} from '../utils.js';
 
 const createBladeCursorRulesToolName = 'create_blade_cursor_rules';
 
@@ -21,47 +26,54 @@ const createBladeCursorRulesToolSchema = {
 const createBladeCursorRulesToolCallback: ToolCallback<
   typeof createBladeCursorRulesToolSchema
 > = async ({ currentProjectRootDirectory }) => {
-  const ruleFileDir = join(currentProjectRootDirectory, '.cursor/rules');
-  const ruleFilePath = join(ruleFileDir, 'frontend-blade-rules.mdc');
+  try {
+    const ruleFileDir = join(currentProjectRootDirectory, '.cursor/rules');
+    const ruleFilePath = join(ruleFileDir, 'frontend-blade-rules.mdc');
 
-  if (existsSync(ruleFilePath)) {
-    if (hasOutDatedRules(ruleFilePath)) {
-      // removes the outdated rules file and continues execution to generate new rule file
-      unlinkSync(ruleFilePath);
-    } else {
-      return {
-        content: [{ type: 'text', text: 'Cursor rules already exist. Doing nothing' }],
-      };
+    if (existsSync(ruleFilePath)) {
+      if (hasOutDatedRules(ruleFilePath)) {
+        // removes the outdated rules file and continues execution to generate new rule file
+        unlinkSync(ruleFilePath);
+      } else {
+        return {
+          content: [{ type: 'text', text: 'Cursor rules already exist. Doing nothing' }],
+        };
+      }
     }
-  }
 
-  const ruleFileTemplateContent = readFileSync(BLADE_CURSOR_RULES_FILE_PATH, 'utf8').replace(
-    'rules_version: <!-- dynamic_version -->',
-    `rules_version: ${CURSOR_RULES_VERSION}`,
-  );
+    const ruleFileTemplateContent = readFileSync(BLADE_CURSOR_RULES_FILE_PATH, 'utf8').replace(
+      'rules_version: <!-- dynamic_version -->',
+      `rules_version: ${CURSOR_RULES_VERSION}`,
+    );
 
-  if (!existsSync(ruleFileDir)) {
-    mkdirSync(ruleFileDir, { recursive: true });
-  }
+    if (!existsSync(ruleFileDir)) {
+      mkdirSync(ruleFileDir, { recursive: true });
+    }
 
-  writeFileSync(ruleFilePath, ruleFileTemplateContent);
+    writeFileSync(ruleFilePath, ruleFileTemplateContent);
 
-  await sendAnalytics({
-    eventName: 'Blade MCP Tool Called',
-    properties: {
-      toolName: createBladeCursorRulesToolName,
-      cursorRulesVersion: CURSOR_RULES_VERSION,
-    },
-  });
-
-  return {
-    content: [
-      {
-        type: 'text',
-        text: `Blade cursor rules created at: ${ruleFilePath}. Cursor Rules Version: ${CURSOR_RULES_VERSION}`,
+    await sendAnalytics({
+      eventName: 'Blade MCP Tool Called',
+      properties: {
+        toolName: createBladeCursorRulesToolName,
+        cursorRulesVersion: CURSOR_RULES_VERSION,
       },
-    ],
-  };
+    });
+
+    return {
+      content: [
+        {
+          type: 'text',
+          text: `Blade cursor rules created at: ${ruleFilePath}. Cursor Rules Version: ${CURSOR_RULES_VERSION}`,
+        },
+      ],
+    };
+  } catch (error: unknown) {
+    return handleError({
+      toolName: createBladeCursorRulesToolName,
+      error,
+    });
+  }
 };
 
 export {
