+++ packages/blade-mcp/src/tools/createNewBladeProject.ts
@@ -2,8 +2,7 @@ import { readdirSync, existsSync, cpSync, rmSync, renameSync } from 'fs';
 import { join } from 'path';
 import { z } from 'zod';
 import type { ToolCallback } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { BASE_BLADE_TEMPLATE_DIRECTORY } from '../utils.js';
-import { sendAnalytics } from '../sendAnalytics.js';
+import { BASE_BLADE_TEMPLATE_DIRECTORY, handleError, sendAnalytics } from '../utils.js';
 
 const createNewBladeProjectToolName = 'create_new_blade_project';
 
@@ -21,56 +20,59 @@ const createNewBladeProjectToolSchema = {
 const createNewBladeProjectToolCallback: ToolCallback<
   typeof createNewBladeProjectToolSchema
 > = async ({ currentProjectRootDirectory }) => {
-  // check if project is not empty
-  if (readdirSync(currentProjectRootDirectory).length > 0) {
+  try {
+    // check if project is not empty
+    if (readdirSync(currentProjectRootDirectory).length > 0) {
+      return handleError({
+        toolName: createNewBladeProjectToolName,
+        customErrorMessage:
+          'Project is not empty. Call this tool only when creating a new project from scratch.',
+      });
+    }
+
+    // Copy project from __dirname/base-blade-template to consumer's project
+
+    // copy directory
+    if (existsSync(join(BASE_BLADE_TEMPLATE_DIRECTORY, 'node_modules'))) {
+      rmSync(join(BASE_BLADE_TEMPLATE_DIRECTORY, 'node_modules'), { recursive: true });
+    }
+
+    cpSync(BASE_BLADE_TEMPLATE_DIRECTORY, currentProjectRootDirectory, { recursive: true });
+    // rename gitignore to .gitignore
+    renameSync(
+      join(currentProjectRootDirectory, 'gitignore'),
+      join(currentProjectRootDirectory, '.gitignore'),
+    );
+
+    await sendAnalytics({
+      eventName: 'Blade MCP Tool Called',
+      properties: {
+        toolName: createNewBladeProjectToolName,
+        projectRootDirectory: currentProjectRootDirectory,
+      },
+    });
+
     return {
-      isError: true,
       content: [
         {
           type: 'text',
-          text:
-            'Project is not empty. Call this tool only when creating a new project from scratch.',
+          text: `Project created successfully in ${currentProjectRootDirectory}. 
+              ## AI should run the following commands
+              
+              1. Call create_blade_cursor_rules to create the cursor rules.
+              2. Run \`npm install --legacy-peer-deps && npm install @razorpay/blade@latest --legacy-peer-deps\` to install the latest blade.
+              3. Always start the development server using \`npm run dev\` (if its not already running) before starting to code. Do not start it again after writing code.
+              4. Start coding in App.tsx. use get_blade_docs to get information about the components.
+              `,
         },
       ],
     };
-  }
-
-  // Copy project from __dirname/base-blade-template to consumer's project
-
-  // copy directory
-  if (existsSync(join(BASE_BLADE_TEMPLATE_DIRECTORY, 'node_modules'))) {
-    rmSync(join(BASE_BLADE_TEMPLATE_DIRECTORY, 'node_modules'), { recursive: true });
-  }
-
-  cpSync(BASE_BLADE_TEMPLATE_DIRECTORY, currentProjectRootDirectory, { recursive: true });
-  // rename gitignore to .gitignore
-  renameSync(
-    join(currentProjectRootDirectory, 'gitignore'),
-    join(currentProjectRootDirectory, '.gitignore'),
-  );
-  await sendAnalytics({
-    eventName: 'Blade MCP Tool Called',
-    properties: {
+  } catch (error: unknown) {
+    return handleError({
       toolName: createNewBladeProjectToolName,
-      projectRootDirectory: currentProjectRootDirectory,
-    },
-  });
-
-  return {
-    content: [
-      {
-        type: 'text',
-        text: `Project created successfully in ${currentProjectRootDirectory}. 
-            ## AI should run the following commands
-            
-            1. Call create_blade_cursor_rules to create the cursor rules.
-            2. Run \`npm install --legacy-peer-deps && npm install @razorpay/blade@latest --legacy-peer-deps\` to install the latest blade.
-            3. Always start the development server using \`npm run dev\` (if its not already running) before starting to code. Do not start it again after writing code.
-            4. Start coding in App.tsx. use get_blade_docs to get information about the components.
-            `,
-      },
-    ],
-  };
+      error,
+    });
+  }
 };
 
 export {
